Move namespace from `JSON::API` to `JSONAPI`.
