--- lib/json/api.rb
@@ -1,20 +1,0 @@
-require 'json/api/exceptions'
-require 'json/api/version'
-
-require 'json/api/attributes'
-require 'json/api/document'
-require 'json/api/error'
-require 'json/api/jsonapi'
-require 'json/api/link'
-require 'json/api/links'
-require 'json/api/relationship'
-require 'json/api/relationships'
-require 'json/api/resource'
-require 'json/api/resource_identifier'
-
-require 'json/api/parse'
-
-module JSON
-  module API
-  end
-end
