--- lib/json/api/exceptions.rb
@@ -1,6 +1,0 @@
-module JSON
-  module API
-    class InvalidDocument < StandardError
-    end
-  end
-end
