--- lib/json/api/version.rb
@@ -1,5 +1,0 @@
-module JSON
-  module API
-    VERSION = '0.5.2'.freeze
-  end
-end
