+++ lib/jsonapi.rb
@@ -1,0 +1,18 @@
+require 'jsonapi/exceptions'
+require 'jsonapi/version'
+
+require 'jsonapi/attributes'
+require 'jsonapi/document'
+require 'jsonapi/error'
+require 'jsonapi/jsonapi'
+require 'jsonapi/link'
+require 'jsonapi/links'
+require 'jsonapi/relationship'
+require 'jsonapi/relationships'
+require 'jsonapi/resource'
+require 'jsonapi/resource_identifier'
+
+require 'jsonapi/parse'
+
+module JSONAPI
+end
