+++ lib/jsonapi/exceptions.rb
@@ -1,0 +1,4 @@
+module JSONAPI
+  class InvalidDocument < StandardError
+  end
+end
