+++ lib/jsonapi/version.rb
@@ -1,0 +1,3 @@
+module JSONAPI
+  VERSION = '0.5.2'.freeze
+end
