+++ spec/include_directive/parser_spec.rb
@@ -1,12 +1,12 @@
-require 'json/api/include_directive'
+require 'jsonapi/include_directive'
 
-describe JSON::API::IncludeDirective::Parser, '.parse_include_args' do
+describe JSONAPI::IncludeDirective::Parser, '.parse_include_args' do
   it 'handles arrays of symbols and hashes' do
     args = [:friends,
             comments: [:author],
             posts: [:author,
                     comments: [:author]]]
-    hash = JSON::API::IncludeDirective::Parser.parse_include_args(args)
+    hash = JSONAPI::IncludeDirective::Parser.parse_include_args(args)
     expected = {
       friends: {},
       comments: { author: {} },
@@ -18,7 +18,7 @@ describe JSON::API::IncludeDirective::Parser, '.parse_include_args' do
 
   it 'handles strings with spaces' do
     str = 'friends, comments.author, posts.author, posts.comments.author'
-    hash = JSON::API::IncludeDirective::Parser.parse_include_args(str)
+    hash = JSONAPI::IncludeDirective::Parser.parse_include_args(str)
     expected = {
       friends: {},
       comments: { author: {} },
@@ -31,7 +31,7 @@ describe JSON::API::IncludeDirective::Parser, '.parse_include_args' do
   it 'handles common prefixes in strings' do
     args = ['friends', 'comments.author', 'posts.author',
             'posts.comments.author']
-    hash = JSON::API::IncludeDirective::Parser.parse_include_args(args)
+    hash = JSONAPI::IncludeDirective::Parser.parse_include_args(args)
     expected = {
       friends: {},
       comments: { author: {} },
@@ -43,7 +43,7 @@ describe JSON::API::IncludeDirective::Parser, '.parse_include_args' do
 
   it 'handles an empty string' do
     args = ''
-    hash = JSON::API::IncludeDirective::Parser.parse_include_args(args)
+    hash = JSONAPI::IncludeDirective::Parser.parse_include_args(args)
     expected = {}
 
     expect(hash).to eq expected
@@ -51,7 +51,7 @@ describe JSON::API::IncludeDirective::Parser, '.parse_include_args' do
 
   it 'handles an empty array' do
     args = []
-    hash = JSON::API::IncludeDirective::Parser.parse_include_args(args)
+    hash = JSONAPI::IncludeDirective::Parser.parse_include_args(args)
     expected = {}
 
     expect(hash).to eq expected
